pyproject.toml
[tool.poetry]
name = "td7"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
apache-airflow = "^2.9.1"
dbt-core = "^1.8.1"
dbt-postgres = "^1.8.1"
Faker = "^25.3.0"
pandas = "=2.0.0"
python-dotenv = "^1.0.1"
astronomer-cosmos = "^1.4.1"
psycopg2 = "^2.9.9"
apache-airflow-providers-postgres = "^5.11.1"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
